+++ packages/core/src/odata/common/filter/filter.ts
@@ -4,7 +4,7 @@ import { Constructable } from '../constructable';
 import { EdmTypeShared } from '../edm-types';
 import { EntityBase, EntityIdentifiable, ODataVersionOf } from '../entity';
 import { FieldType } from '../selectable';
-import { FilterFunction } from './filter-function';
+import { FilterFunction } from './filter-function-base';
 import { Filterable } from './filterable';
 
 type FilterOperatorString = 'eq' | 'ne';
