+++ packages/core/src/odata/common/filter/index.ts
@@ -3,6 +3,7 @@
 export * from './boolean-filter-function';
 export * from './filter';
 export * from './filter-function';
+export * from './filter-function-base';
 export * from './filter-link';
 export * from './filter-list';
 export * from './filterable';
